ufl.Function().geometric_dimension() deprecated
In the latest version of dolfinx (currently v0.5.2 released on Oct 27) includes the latest version of UFL(v2022.2.0) no longer suport the method geometric_dimension of the ufl.Function. The exact function can be imported with: `from ufl.domain import find_geometric_dimension` but in the dolfinx demos for python the prerrefered way is using len().